Modified Specials, Testimonial and About package
In specials three cards were created.
In the Testimonial page an array was created to hold the name,
image and the comment of the termonial. useState was created
currentindex and to setindex were passed with a default value of 0;
two buttons were created to navigate through the slide, on each button
a function prevSlide or nextSlide is called of which the setindex is also
called and a tenary operator to check if its the end of the array then currentindex
should set to zero else setindex should increase currentindex by 1;