Change Available Times State to a useReducer
Created a component called MyContext in which availableTimes useReducer
was create, and a function updateTimes was passed as
parameter that will be performed on the state, to get a new state
to get a new state and initialTimes was also passed
as the initial value of the availableTimes. availableTimes
and dispatch was now passed to other components as children.
On the BookingForm a useEffect was Created
to handle the dispatch based on the newTime and
the selected date.